Update ordinary-town.md
I specified that if you are town you won't see any channels in the special-role-private-channels folder. I also wrote that if someone would happen to be unsure of their role they should contact the game masters. But I don't know if that is unnecessary or if it should be stated in general rules or something instead.  Also, do you think it should be written somewhere in the rules that unless a player is sure how to play the game it is recommended to spectate first game? I didn't my first game and i think that it's good that it is not a rule that you have to spectate first round. But especially with these new rules it might not be needed, as long as new players have read through the rules carefully. What do you think? Sorry for the long message :(